+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add `Array::subset_all()`
+
 ## [0.17.0-beta.2] - 2024-09-23
 
 ### Highlights / Major Changes